Move Layout class to Responses namespace
