+++ CopticChanter.WebApi.Client/CoptClient.cs
@@ -1,5 +1,6 @@
 ﻿using System.Xml.Linq;
 using CopticChanter.WebApi.Core;
+using CopticChanter.WebApi.Core.Responses;
 using CoptLib.Writing;
 using Flurl;
 using Flurl.Http;
